Added the Event class for fine-grain control of event dispatch
For a more fine-grain control of event dispatch, the Event class
can be manually instantiated and configured. An `Event` represents
an event as a C++ style function object and can be directly passed
to other APIs that expect a callback.

More information and examples can be found in the updated readme
and documentation in Event.h